Add retain flag to MQTT messages
Every MQTT message is now flagged with the `retain` label which will
ensure that new clients are able to pick up previous messages that are
e.g. only published on the first startup or very irregularly.

The only topic that is not flagged is is the one for the cleaning logs
as this will, over time, lead to a big amount of memory being allocated
on the MQTT broker.